+++ src/entries.test.ts
@@ -21,6 +21,7 @@ import * as v25 from "./v25";
 import * as v26 from "./v26";
 import * as v27 from "./v27";
 import * as v28 from "./v28";
+import * as v29 from "./v29";
 import * as v9 from "./v9";
 
 /**
@@ -52,6 +53,7 @@ const ENTRIES: Record<number, { BYPP_FORMAT_VERSION: number }> = {
   26: v26,
   27: v27,
   28: v28,
+  29: v29,
 };
 
 const SRC = dirname(fileURLToPath(import.meta.url));