+++ tsup.config.ts
@@ -23,6 +23,7 @@ export default defineConfig({
     "src/v26.ts",
     "src/v27.ts",
     "src/v28.ts",
+    "src/v29.ts",
     "src/icons.ts",
   ],
   format: ["esm", "cjs"],